Migrate AfterPay internal components to TypeScript, improve code structure and add config file.
