+++ src/components/AfterPay/components/ConsentCheckbox/index.ts
